backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pywin32==306
requests==2.32.3
lxml==5.2.2
cryptography==43.0.1
python-dotenv==1.0.1
